+++ src/lambdacd/pipeline_state.clj
@@ -66,25 +66,5 @@
 (defn running [ctx]
   (update ctx {:status :running}))
 
-(defn- status-for-steps [steps]
-  (let [statuses (map :status (vals steps))
-        has-failed (util/contains-value? :failure statuses)
-        has-running (util/contains-value? :running statuses)
-        has-waiting (util/contains-value? :waiting statuses)
-        all-ok (every? #(= % :ok) statuses)]
-    (cond
-      has-failed :failure
-      has-running :running
-      all-ok :ok
-      has-waiting :waiting
-      :else :unknown)))
-
-(defn- history-entry [[k v]]
-  { :build-number k
-    :status (status-for-steps v)})
-
-(defn history-for [state]
-  (map history-entry state))
-
 (defn most-recent-build-number-in [state]
   (apply max (keys state)))